now shuts down when no SD is detected -- Annon contrib.
