--- payload_stage1/source/hid.c
@@ -1,15 +1,0 @@
-#include "hid.h"
-
-u32 InputWait() {
-    u32 pad_state_old = HID_STATE;
-    while (true) {
-        u32 pad_state = HID_STATE;
-        if (pad_state ^ pad_state_old)
-            return ~pad_state;
-    }
-}
-
-unsigned int GetInput()
-{
-	return ~HID_STATE;
-}
